Move _split_lineage into spine
Pure refactor. The index allocator landing next lives in spine.py and
needs to peel a Lineage off a filename, and the import direction runs
survey -> spine, so the helper has to move rather than be imported.
